+++ tests/testthat/test-l_cols_align.R
@@ -11,6 +11,7 @@ test_that("cols_align() works correctly", {
   tbl_latex <-
     gt(mtcars_short) %>%
     cols_align(align = "left", columns = c(mpg, cyl, drat)) %>%
+    tab_options(latex.use_longtable = TRUE) %>%
     as_latex() %>% as.character()
 
   # Expect a characteristic pattern
@@ -22,6 +23,7 @@ test_that("cols_align() works correctly", {
   tbl_latex <-
     gt(mtcars_short) %>%
     cols_align(align = "left", columns = 1:3) %>%
+    tab_options(latex.use_longtable = TRUE) %>%
     as_latex() %>% as.character()
 
   # Expect a characteristic pattern
@@ -33,6 +35,7 @@ test_that("cols_align() works correctly", {
   tbl_latex <-
     gt(mtcars_short) %>%
     cols_align(align = "left") %>%
+    tab_options(latex.use_longtable = TRUE) %>%
     as_latex() %>% as.character()
 
   # Expect a characteristic pattern
@@ -44,6 +47,7 @@ test_that("cols_align() works correctly", {
   tbl_latex <-
     gt(mtcars_short) %>%
     cols_align(align = "left", columns = everything()) %>%
+    tab_options(latex.use_longtable = TRUE) %>%
     as_latex() %>% as.character()
 
   # Expect a characteristic pattern
@@ -55,6 +59,7 @@ test_that("cols_align() works correctly", {
   tbl_latex <-
     gt(sp500) %>%
     cols_align(align = "auto") %>%
+    tab_options(latex.use_longtable = TRUE) %>%
     as_latex() %>% as.character()
 
   # Expect a characteristic pattern